Fourth task is done

+++ src/utils.h
@@ -5,6 +5,6 @@
 #ifndef FIELDS_OF_GOLD_UTILS_H
 #define FIELDS_OF_GOLD_UTILS_H
 
-using numeric_type = double;
+using numeric_type = long double;
 
 #endif //FIELDS_OF_GOLD_UTILS_H
